Finish the cmd/render.go per-command split
Second (and final) installment on structural refactor #9. The previous
PR (#322) extracted the four largest per-command renderers. This PR
finishes the split for every remaining command and collapses cmd/render.go
to the dispatcher layer alone.

cmd/render.go: 1322 -> 127 LOC (-90%).
Net across both PRs: 2068 -> 127 LOC (-94%).

New per-command files (each with its primary render function plus any
command-local helpers):

- cmd/render_index.go        (index:     RebuildResult, reuse summary,
                              repo coverage, governance delta, source
                              summaries)
- cmd/render_discover.go     (discover + canonicalize + new:
                              DiscoverResult, CanonicalizeResult,
                              NewSpecBundleResult — grouped because they
                              all produce draft workspace artifacts)
- cmd/render_init.go         (init + migrate-config + version:
                              InitResult, MigrateConfigResult,
                              VersionResult — setup/lifecycle outputs)
- cmd/render_preview.go      (preview-sources)
- cmd/render_explain.go      (explain-file)
- cmd/render_search.go       (search-specs: text + table renderers and
                              score helpers)
- cmd/render_overlap.go      (check-overlap + compare-specs + overlap
                              block/badge/humanizer helpers)
- cmd/render_terminology.go  (check-terminology + finding renderer)
- cmd/render_freshness.go    (check-spec-freshness + verdict label)
- cmd/render_fix.go          (fix + compile + fix-prompt-file)

Updated existing files:

- cmd/render_review.go: now owns the top-level renderReviewResult,
  renderReviewMarkdown, renderReviewHTML entry points plus the
  review-specific helpers (reviewHTMLDriftEvidence, reviewHTMLRemediationEvidence,
  reviewMarkdownSummary, reviewMarkdownActions, renderReviewMarkdownDriftEvidence)
  alongside the Markdown/HTML section renderers that already lived there.

- cmd/presentation.go: inherited the treeBranch/treeItem methods on
  renderPresentation (they were defined in render.go but read more
  naturally next to the sibling treeMid/treeLast methods).

- cmd/render_util.go (new): shared rendering helpers used by 2+
  command renderers — renderFreshnessLabel, statusEmbedderSummary,
  preferredDocLabel, displaySourcePath, displayDefault, repoPathLabel,
  isNonPrimaryRepoDoc, pluralSuffix, humanizeSymbol, truncateRenderLine,
  minInt.

- cmd/render.go: now a 127-LOC dispatcher file containing only the
  four top-level entry points (renderCommandResult,
  renderCommandTable, renderCommandMarkdown, renderCommandHTML) and
  usesSemanticTextRendering.

Every function is moved verbatim. No signature or logic changes, no new
abstraction (the review's suggested sectionRenderer primitive is a
separate design exercise, not in scope here). All 18 existing render
tests pass without modification. Verified with make fmt, make vet,
go test ./..., and go test -race ./cmd/... .
