+++ cmd/render_discover.go
@@ -1,0 +1,79 @@
+package cmd
+
+import (
+	"fmt"
+	"io"
+	"strings"
+
+	"github.com/dusk-network/pituitary/internal/source"
+)
+
+func renderDiscoverResult(w io.Writer, result *source.DiscoverResult) {
+	fmt.Fprintf(w, "workspace: %s\n", result.WorkspaceRoot)
+	fmt.Fprintf(w, "config path: %s\n", result.ConfigPath)
+	if result.WroteConfig {
+		fmt.Fprintln(w, "config write: wrote local config")
+	} else {
+		fmt.Fprintln(w, "config write: skipped")
+	}
+
+	for _, discovered := range result.Sources {
+		fmt.Fprintf(w, "source: %s | %s | root: %s | items: %d | confidence: %s\n", discovered.Name, discovered.Kind, discovered.Path, discovered.ItemCount, discovered.Confidence)
+		for _, reason := range discovered.Rationale {
+			fmt.Fprintf(w, "rationale: %s\n", reason)
+		}
+		for _, item := range discovered.Items {
+			fmt.Fprintf(w, "- %s | %s\n", item.Path, item.Confidence)
+		}
+	}
+
+	if result.Preview != nil {
+		fmt.Fprintf(w, "preview sources: %d\n", len(result.Preview.Sources))
+	}
+
+	fmt.Fprintln(w, "generated config:")
+	fmt.Fprint(w, result.Config)
+}
+
+func renderCanonicalizeResult(w io.Writer, result *source.CanonicalizeResult) {
+	fmt.Fprintf(w, "workspace: %s\n", result.WorkspaceRoot)
+	fmt.Fprintf(w, "source: %s\n", result.SourcePath)
+	fmt.Fprintf(w, "bundle dir: %s\n", result.BundleDir)
+	if result.WroteBundle {
+		fmt.Fprintln(w, "bundle write: wrote generated bundle")
+	} else {
+		fmt.Fprintln(w, "bundle write: skipped")
+	}
+	fmt.Fprintf(w, "spec ref: %s\n", result.Spec.Ref)
+	fmt.Fprintf(w, "title: %s\n", result.Spec.Title)
+	if result.Spec.Inference != nil {
+		fmt.Fprintf(w, "inference: %s (%.2f)\n", result.Spec.Inference.Level, result.Spec.Inference.Score)
+	}
+	fmt.Fprintf(w, "provenance: %s\n", result.Provenance.SourceRef)
+	for _, file := range result.Files {
+		fmt.Fprintf(w, "generated file: %s\n", file.Path)
+		fmt.Fprint(w, file.Content)
+		if !strings.HasSuffix(file.Content, "\n") {
+			fmt.Fprintln(w)
+		}
+	}
+}
+
+func renderNewSpecBundleResult(w io.Writer, result *source.NewSpecBundleResult) {
+	fmt.Fprintf(w, "workspace: %s\n", result.WorkspaceRoot)
+	if result.ConfigPath != "" {
+		fmt.Fprintf(w, "config path: %s\n", result.ConfigPath)
+	}
+	fmt.Fprintf(w, "spec root: %s\n", result.SpecRoot)
+	fmt.Fprintf(w, "bundle dir: %s\n", result.BundleDir)
+	if result.WroteBundle {
+		fmt.Fprintln(w, "bundle write: wrote draft bundle")
+	}
+	fmt.Fprintf(w, "spec ref: %s\n", result.Spec.Ref)
+	fmt.Fprintf(w, "title: %s\n", result.Spec.Title)
+	fmt.Fprintf(w, "status: %s\n", result.Spec.Status)
+	fmt.Fprintf(w, "domain: %s\n", result.Spec.Domain)
+	for _, file := range result.Files {
+		fmt.Fprintf(w, "generated file: %s\n", file.Path)
+	}
+}
